requirements.txt
beautifulsoup4~=4.9.3
bs4==0.0.1
cfscrape==2.1.1
requests==2.28.1
stem~=1.8.0
mysql.connector